+++ src/app/page.js
@@ -1,0 +1,5 @@
+import Image from "next/image"
+
+export default function Home() {
+  return <main>educhain</main>
+}
